cogs/management に admin.pyとeval.pyを移動
