--- cogs/eval.py
@@ -1,81 +1,0 @@
-import asyncio
-import io
-import subprocess
-import textwrap
-import traceback
-from contextlib import redirect_stdout
-
-from discord.ext import commands
-
-
-def cleanup_code(content):
-    if content.startswith('```') and content.endswith('```'):
-        return '\n'.join(content.split('\n')[1:-1])
-    return content.strip('` \n')
-
-
-class Owner(commands.Cog):
-
-    def __init__(self, bot):
-        self.bot = bot
-        self._last_result = None
-        self.sessions = set()
-
-    async def cog_check(self, ctx):
-        return await self.bot.is_owner(ctx.author)
-
-    async def run_process(self, command):
-        try:
-            process = await asyncio.create_subprocess_shell(
-                command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-            result = await process.communicate()
-        except NotImplementedError:
-            process = subprocess.Popen(
-                command, shell=True, stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE)
-            result = await self.bot.loop.run_in_executor(
-                None, process.communicate)
-        return [output.decode() for output in result]
-
-    @commands.command(pass_context=True, hidden=True)
-    async def eval(self, ctx, *, body: str):
-        env = {
-            'bot': self.bot,
-            'ctx': ctx,
-            'channel': ctx.channel,
-            'author': ctx.author,
-            'guild': ctx.guild,
-            'message': ctx.message,
-            '_': self._last_result
-        }
-        env.update(globals())
-        body = cleanup_code(body)
-        stdout = io.StringIO()
-        to_compile = f'async def func():\n{textwrap.indent(body, "  ")}'
-        try:
-            exec(to_compile, env)
-        except Exception as e:
-            return await ctx.send(f'```py\n{e.__class__.__name__}: {e}\n```')
-        func = env['func']
-        try:
-            with redirect_stdout(stdout):
-                ret = await func()
-        except Exception:
-            value = stdout.getvalue()
-            await ctx.send(f'```py\n{value}{traceback.format_exc()}\n```')
-        else:
-            value = stdout.getvalue()
-            try:
-                await ctx.message.add_reaction('\u2705')
-            except Exception:
-                pass
-            if ret is None:
-                if value:
-                    await ctx.send(f'```py\n{value}\n```')
-            else:
-                self._last_result = ret
-                await ctx.send(f'```py\n{value}{ret}\n```')
-
-
-def setup(bot):
-    bot.add_cog(Owner(bot))
